Last update

+++ server/index.js
@@ -20,6 +20,14 @@ const path = require('path');
 app.get('/', (req, res)=> res.sendFile(path.join(__dirname, '../client/dist/index.html')));
 app.use('/assets', express.static(path.join(__dirname, '../client/dist/assets'))); 
 
+const isLoggedIn = async(req, res, next) => {
+  try {
+    req.user = await findUserWithToken(req.headers.authorization);
+    next();
+  } catch (error) {
+    next(error)
+  }
+}
 
 app.post('/api/auth/login', async(req, res, next)=> {
   try {
@@ -30,9 +38,9 @@ app.post('/api/auth/login', async(req, res, next)=> {
   }
 });
 
-app.get('/api/auth/me', async(req, res, next)=> {
+app.get('/api/auth/me', isLoggedIn, (req, res, next)=> {
   try {
-    res.send(await findUserWithToken(req.headers.authorization));
+    res.send(req.user);
   }
   catch(ex){
     next(ex);
@@ -48,8 +56,13 @@ app.get('/api/users', async(req, res, next)=> {
   }
 });
 
-app.get('/api/users/:id/favorites', async(req, res, next)=> {
+app.get('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
   try {
+    if (req.params.id !== req.user.id) {
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     res.send(await fetchFavorites(req.params.id));
   }
   catch(ex){
@@ -57,8 +70,21 @@ app.get('/api/users/:id/favorites', async(req, res, next)=> {
   }
 });
 
-app.post('/api/users/:id/favorites', async(req, res, next)=> {
+app.post('/api/auth/register', async(req, res, next) =>{
   try {
+    res.send(await createUser(req.body))
+  } catch (error) {
+    next(error)
+  }
+})
+
+app.post('/api/users/:id/favorites', isLoggedIn, async(req, res, next)=> {
+  try {
+    if (req.params.id !== req.user.id) {
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     res.status(201).send(await createFavorite({ user_id: req.params.id, product_id: req.body.product_id}));
   }
   catch(ex){
@@ -66,8 +92,13 @@ app.post('/api/users/:id/favorites', async(req, res, next)=> {
   }
 });
 
-app.delete('/api/users/:user_id/favorites/:id', async(req, res, next)=> {
+app.delete('/api/users/:user_id/favorites/:id', isLoggedIn, async(req, res, next)=> {
   try {
+    if (req.params.id !== req.user.id) {
+      const error = Error('not authorized');
+      error.status = 401;
+      throw error;
+    }
     await destroyFavorite({user_id: req.params.user_id, id: req.params.id });
     res.sendStatus(204);
   }
